+++ src/lib/store.ts
@@ -11,265 +11,15 @@ import { notify } from "./notify";
 import { playArrival, playDone, playError } from "./sound";
 import { autoUpdateEnabled, checkForUpdatesQuietly } from "./updates";
 import { DEFAULT_CHAT_CONFIG, DEFAULT_READING_PREFS } from "./types";
+import type { AppState, Migration, QueueItem } from "./storeTypes";
+export type { ExternalAdd, Migration, QueueItem } from "./storeTypes";
 import type {
-  AiConfig,
   ChatConfig,
-  KokoroStatus,
   Message,
-  ModelHealth,
-  ModelStat,
   Note,
-  NoteKind,
-  Notebook,
   ReadingPrefs,
   ReportSchedule,
-  Source,
-  Template,
-  Toast,
-  ToastKind,
 } from "./types";
-
-export interface QueueItem {
-  id: string;
-  name: string;
-  status: "pending" | "processing" | "done" | "error";
-  error?: string;
-}
-
-/** What an external workflow (deep link, Services, tray) wants to add. */
-export interface ExternalAdd {
-  files: string[];
-  url: string | null;
-  text: string | null;
-  title: string | null;
-}
-
-export interface Migration {
-  done: number;
-  total: number;
-  title: string;
-}
-
-interface AppState {
-  notebooks: Notebook[];
-  currentId: string | null;
-  sources: Source[];
-  /** Which sources feed chat retrieval and Studio generation. null = all
-   *  selected (the default; new sources are auto-included). A non-null map
-   *  holds ONLY deselected ids, set to false. Persisted per notebook. */
-  selectedSourceIds: Record<string, boolean> | null;
-  messages: Message[];
-  notes: Note[];
-  reportSchedules: ReportSchedule[];
-  /** Custom generators from ~/Documents/Alchemy/templates (global, not per-notebook). */
-  templates: Template[];
-  aiConfig: AiConfig | null;
-  ollamaOk: boolean | null;
-  modelHealth: ModelHealth | null;
-  modelStats: ModelStat[];
-  theme: string;
-  reading: ReadingPrefs;
-
-  sending: boolean;
-  streamingText: string;
-  steps: string[];
-  agentMode: boolean;
-  chatConfig: ChatConfig;
-  followups: string[];
-  summary: string;
-  summaryLoading: boolean;
-  generatingKind: NoteKind | null;
-  /** Which template tile is generating (kind alone can't tell them apart). */
-  generatingTemplateId: string | null;
-  ingestQueue: QueueItem[];
-  migration: Migration | null;
-  draggingFiles: boolean;
-  sourcesOpen: boolean;
-  studioOpen: boolean;
-  /** Draggable side-panel widths (px), persisted. */
-  sourcesWidth: number;
-  studioWidth: number;
-  onboardingDismissed: boolean;
-  settingsOpen: boolean;
-  settingsTab: string;
-  /** Cmd+K command menu. */
-  paletteOpen: boolean;
-  /** The add-source modal (the single "add sources" surface). */
-  addSourceOpen: boolean;
-  /** Deep-link straight to a form; null opens on the hub of tiles. */
-  addSourceStep: "url" | "text" | null;
-  /** Mac providers (Calendar, Reminders, Notes via cider) reachable?
-   *  null = probe still in flight; false shows the install hint instead. */
-  macAvailable: boolean | null;
-  /** Ask for the URL form — a flag (not an event) because the modal may
-   *  still be mounting when the ask happens. */
-  pendingAddUrl: boolean;
-  /** Same, for the paste-text form. */
-  pendingAddText: boolean;
-  /** Menu asked for an update check — Settings' General tab runs it on mount. */
-  pendingUpdateCheck: boolean;
-  /** One-time model download overlay (embedder or TTS voice model). */
-  embedderDownload: {
-    label: string;
-    done: number;
-    total: number;
-    title?: string;
-  } | null;
-  error: string | null;
-  /** Text of a chat send that failed, handed back to the composer so it isn't lost. */
-  failedInput: string | null;
-  /** Text another surface (e.g. the source reader) wants in the composer. */
-  pendingInput: string | null;
-  /** A question the palette should ask across all notebooks on next open
-   *  (the homepage's unified ask box seeds this, then opens the palette). */
-  pendingAsk: string | null;
-  /** Ephemeral toasts (success/info auto-dismiss; errors linger a bit longer). */
-  toasts: Toast[];
-  /** Id of a just-generated note, so the Studio panel can auto-open it. */
-  justCreatedNoteId: string | null;
-  /** Cmd+N pressed while Studio was collapsed — open the composer on mount. */
-  pendingNewNote: boolean;
-  /** Streaming buffer for the in-flight Studio generation (artifact://token). */
-  artifactStreamText: string;
-  /** Audio Overview synthesis progress (audio://progress), null when idle. */
-  audioProgress: { done: number; total: number } | null;
-  /** Podcast voice model readiness; the generator hides until verified. */
-  kokoroStatus: KokoroStatus | null;
-  /** setup_kokoro in flight (Settings → Models). */
-  kokoroBusy: boolean;
-  /** Source open in the reader, optionally scrolled to a cited passage. */
-  viewingSource: { sourceId: string; title: string; highlight?: string } | null;
-  /** Live folder-scan progress (folder://progress), null when idle. */
-  folderScan: { done: number; total: number; title: string } | null;
-  /** Note read-state (notes & reports): note id -> last-read millis.
-   *  Persisted; shared by the home feed, notebook dots, and Studio. */
-  noteReads: Record<string, number>;
-  /** Implicit read horizon — notes older than this predate read tracking
-   *  and count as read, so a fresh install isn't a wall of dots. */
-  noteReadsBaseline: number;
-
-  init: () => Promise<void>;
-  /** Register page-lifetime Tauri event listeners (called once from init). */
-  bindGlobalListeners: () => void;
-  refreshNotebooks: () => Promise<void>;
-  selectNotebook: (id: string) => Promise<void>;
-  closeNotebook: () => void;
-  createNotebook: (title: string) => Promise<void>;
-  renameNotebook: (id: string, title: string) => Promise<void>;
-  setNotebookColor: (id: string, color: string) => Promise<void>;
-  deleteNotebook: (id: string) => Promise<void>;
-  setTheme: (theme: string) => void;
-  setReading: (patch: Partial<ReadingPrefs>) => void;
-  clearQueueItem: (id: string) => void;
-  setDraggingFiles: (v: boolean) => void;
-  toggleSources: () => void;
-  toggleStudio: () => void;
-  setPanelWidth: (panel: "sources" | "studio", width: number) => void;
-  dismissOnboarding: () => void;
-  openSettings: (tab?: string) => void;
-  closeSettings: () => void;
-  setPaletteOpen: (open: boolean) => void;
-  /** Open/close the command menu, refusing to stack over an open dialog. */
-  togglePalette: () => void;
-  /** Open the add-source modal, optionally straight on a form. */
-  openAddSource: (step?: "url" | "text") => void;
-  closeAddSource: () => void;
-  /** Pick a directory and export the current notebook as an OKF bundle. */
-  exportNotebookOkf: () => Promise<void>;
-  /** Save the current notebook as one shareable .okf.zip file. */
-  shareNotebookOkf: () => Promise<void>;
-  /** The import dialog (new notebook vs merge into existing). */
-  importOkfOpen: boolean;
-  /** A dropped bundle path the import dialog should offer directly. */
-  pendingImportPath: string | null;
-  /** Import a bundle (.okf.zip or folder); null notebookId = new notebook. */
-  importOkf: (path: string, notebookId?: string | null) => Promise<void>;
-  createReport: (
-    name: string,
-    kind: string,
-    prompt: string,
-    intervalSecs: number,
-  ) => Promise<void>;
-  updateReport: (r: ReportSchedule) => Promise<void>;
-  deleteReport: (id: string) => Promise<void>;
-  runReportNow: (id: string) => Promise<void>;
-  startReportScheduler: () => void;
-
-  pickAndAddFiles: () => Promise<void>;
-  pickAndAddFolder: () => Promise<void>;
-  addSourceFiles: (paths: string[]) => Promise<void>;
-  /** Start the once-a-minute source resync loop — folder rescans plus
-   *  changed-on-disk re-embeds for loose file sources (main window only). */
-  startSourceSync: () => void;
-  addSourceUrl: (url: string) => Promise<void>;
-  addSourceText: (title: string, text: string) => Promise<void>;
-  addSourceMac: (
-    provider: string,
-    collection: string,
-    label: string,
-  ) => Promise<void>;
-  editSourceText: (
-    sourceId: string,
-    title: string,
-    text: string,
-  ) => Promise<void>;
-  refreshSource: (sourceId: string) => Promise<void>;
-  /** Route an alchemy:// URL (deep link, tray, Services, Spotlight). */
-  handleIntegrationUrl: (raw: string) => Promise<void>;
-  /** External add waiting on a notebook choice (deep link/Services/tray). */
-  pendingExternalAdd: ExternalAdd | null;
-  confirmExternalAdd: (
-    notebookId: string,
-    payload?: ExternalAdd,
-  ) => Promise<void>;
-  updateMacNote: (sourceId: string, body: string) => Promise<void>;
-  addMacReminder: (
-    sourceId: string,
-    title: string,
-    notes?: string,
-  ) => Promise<void>;
-  deleteSource: (id: string) => Promise<void>;
-  toggleSourceSelected: (id: string) => void;
-  setAllSourcesSelected: (selected: boolean) => void;
-
-  sendMessage: (content: string) => Promise<void>;
-  cancelGeneration: (scope?: "chat" | "artifact") => void;
-  openSourceViewer: (
-    sourceId: string,
-    title: string,
-    highlight?: string,
-  ) => void;
-  closeSourceViewer: () => void;
-  appendToken: (t: string) => void;
-  appendStep: (label: string) => void;
-  toggleAgentMode: () => void;
-  setChatConfig: (config: ChatConfig) => void;
-  loadFollowups: () => Promise<void>;
-  refreshSummary: () => Promise<void>;
-  clearChat: () => Promise<void>;
-
-  generateArtifact: (kind: NoteKind, prompt?: string) => Promise<void>;
-  /** Run a user template's instruction through the custom-prompt generation path. */
-  generateFromTemplate: (t: Template) => Promise<void>;
-  rebuildNote: (note: Note) => Promise<void>;
-  createNote: (title: string, content: string) => Promise<void>;
-  updateNote: (id: string, title: string, content: string) => Promise<void>;
-  deleteNote: (id: string) => Promise<void>;
-  discussNoteInChat: (id: string) => Promise<void>;
-  convertNoteToSource: (id: string) => Promise<void>;
-
-  saveAiConfig: (config: AiConfig) => Promise<void>;
-  refreshModelHealth: () => Promise<void>;
-  refreshModelStats: () => Promise<void>;
-  reembedAll: () => Promise<void>;
-  refreshKokoroStatus: () => Promise<void>;
-  setupKokoro: () => Promise<void>;
-  removeKokoro: () => Promise<void>;
-  setError: (e: string | null) => void;
-  pushToast: (kind: ToastKind, message: string) => void;
-  dismissToast: (id: string) => void;
-  markNotesRead: (ids: string[]) => void;
-}
 
 // Side panels stay usable at any drag position: wide enough for content,
 // narrow enough to leave the chat column room at the 1040px minimum window.
